+++ tests/test_broadcast.py
@@ -12,6 +12,8 @@ from frequenz.channels import (
     Broadcast,
     ChannelClosedError,
     Receiver,
+    ReceiverError,
+    ReceiverStoppedError,
     Sender,
     SenderError,
 )
@@ -36,11 +38,13 @@ async def test_broadcast() -> None:
         for ctr in range(num_receivers):
             await chan.send(ctr + 1)
 
-    async def update_tracker_on_receive(receiver_id: int, chan: Receiver[int]) -> None:
+    async def update_tracker_on_receive(receiver_id: int, recv: Receiver[int]) -> None:
         while True:
             try:
-                msg = await chan.receive()
-            except ChannelClosedError:
+                msg = await recv.receive()
+            except ReceiverStoppedError as err:
+                assert err.receiver is recv
+                assert isinstance(err.__cause__, ChannelClosedError)
                 return
             recv_trackers[receiver_id] += msg
 
@@ -76,8 +80,11 @@ async def test_broadcast_after_close() -> None:
 
     with pytest.raises(SenderError):
         await sender.send(5)
-    with pytest.raises(ChannelClosedError):
+    with pytest.raises(ReceiverStoppedError) as excinfo:
         await receiver.receive()
+    assert excinfo.value.receiver is receiver
+    assert isinstance(excinfo.value.__cause__, ChannelClosedError)
+    assert excinfo.value.__cause__.channel is bcast
 
 
 async def test_broadcast_overflow() -> None:
@@ -171,7 +178,7 @@ async def test_broadcast_peek() -> None:
     peekable = receiver.into_peekable()
     sender = bcast.new_sender()
 
-    with pytest.raises(EOFError):
+    with pytest.raises(ReceiverError, match=r"This receiver is no longer active."):
         await receiver.receive()
 
     assert peekable.peek() is None